:recycle::package::wrench: Create environment variable TEST_RUN_MIGRATIONS_AND_SEEDERS, execute migrations and seeders if TEST_RUN_MIGRATIONS_AND_SEEDERS is true and added package postgres.

+++ src/database/migrations/20200518030114-create-positions.js
@@ -1,38 +1,46 @@
 'use strict';
 module.exports = {
   up: (queryInterface, Sequelize) => {
-    return queryInterface.createTable('positions', {
-      id: {
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-        type: Sequelize.INTEGER
-      },
-      name: {
-        allowNull: false,
-        type: Sequelize.STRING,
-      },
-      status: {
-        allowNull: false,
-        type: Sequelize.ENUM('active', 'inactive'),
-        defaultValue: 'active'
-      },
-      max_hired: {
-        allowNull: false,
-        defaultValue: 0,
-        type: Sequelize.INTEGER
-      },
-      created_at: {
-        allowNull: false,
-        type: Sequelize.DATE
-      },
-      updated_at: {
-        allowNull: false,
-        type: Sequelize.DATE
-      }
-    });
+    if (process.env.TEST_RUN_MIGRATIONS_AND_SEEDERS === 'true') {
+      return queryInterface.createTable('positions', {
+        id: {
+          allowNull: false,
+          autoIncrement: true,
+          primaryKey: true,
+          type: Sequelize.INTEGER
+        },
+        name: {
+          allowNull: false,
+          type: Sequelize.STRING,
+        },
+        status: {
+          allowNull: false,
+          type: Sequelize.ENUM('active', 'inactive'),
+          defaultValue: 'active'
+        },
+        max_hired: {
+          allowNull: false,
+          defaultValue: 0,
+          type: Sequelize.INTEGER
+        },
+        created_at: {
+          allowNull: false,
+          type: Sequelize.DATE
+        },
+        updated_at: {
+          allowNull: false,
+          type: Sequelize.DATE
+        }
+      });
+    } else {
+      return Promise.resolve(null);
+    }
   },
   down: (queryInterface, Sequelize) => {
-    return queryInterface.dropTable('positions');
+    if (process.env.TEST_RUN_MIGRATIONS_AND_SEEDERS === 'true') {
+      return queryInterface.dropTable('positions');
+    } else {
+      return Promise.resolve(null);
+    }
   }
 };
